feat: registro de alterações de compra e perda no estoque

+++ src/main/java/com/pub/domain/model/UnidadeConversao.java
@@ -28,7 +28,7 @@ public class UnidadeConversao {
     @JoinColumn(name = "unidade_destino")
 	private Unidade unidadeDestino;
 	
-	private Double fatorConversao;
+	private Integer fatorConversao;
 
 	@Override
 	public int hashCode() {
